избавляется от моков оффера и мест неподалеку
