--- src/mocks/offers-nearby.ts
@@ -1,83 +1,0 @@
-import { CityLocation, Location } from '../types/offers';
-
-export type OfferNearby = {
-  id: string;
-  title: string;
-  type: string;
-  price: number;
-  city: CityLocation;
-  location: Location;
-  isFavorite: boolean;
-  isPremium: boolean;
-  rating: number;
-  previewImage: string;
-}
-
-export const offersNearby: OfferNearby[] = [
-  {
-    'id': '6af6f711-c28d-4121-82cd-e0b462a27f00',
-    'title': 'Beautiful & luxurious studio at great location',
-    'type': 'apartment',
-    'price': 120,
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.35514938496378,
-        'longitude': 4.673877537499948,
-        'zoom': 8
-      }},
-    'location': {
-      'latitude': 52.3909553943508,
-      'longitude': 4.85309666406198,
-      'zoom': 8
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/2.jpg'
-  },
-  {
-    'id': '6af6f711-c28d-4121-82cd-e0b462a27f00',
-    'title': 'Beautiful & luxurious studio at great location',
-    'type': 'apartment',
-    'price': 120,
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.35514938496378,
-        'longitude': 4.673877537499948,
-        'zoom': 8
-      }},
-    'location': {
-      'latitude': 52.3609553943508,
-      'longitude': 4.85309666406198,
-      'zoom': 8
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/19.jpg'
-  },
-  {
-    'id': '6af6f711-c28d-4121-82cd-e0b462a27f00',
-    'title': 'Beautiful & luxurious studio at great location',
-    'type': 'apartment',
-    'price': 120,
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.35514938496378,
-        'longitude': 4.673877537499948,
-        'zoom': 8
-      }},
-    'location': {
-      'latitude': 52.35514938496378,
-      'longitude': 4.673877537499948,
-      'zoom': 8
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/8.jpg'
-  },
-];
