+++ src/pages/offer/offer.tsx
@@ -6,7 +6,6 @@ import { useAppDispatch, useAppSelector } from '../../hooks';
 import { fetchSelectOffer, fetchOffersNearby } from '../../store/api-actions';
 import { dropOffer } from '../../store/action';
 import { selectedOffer } from '../../store/selectors';
-import { offer } from '../../mocks/offer';
 import OfferGoods from './components/offer-goods';
 import ReviewsList from './components/reviews-list';
 import Map from '../../components/map';
@@ -35,7 +34,7 @@ function Offer ({reviews}: OfferProps): JSX.Element {
 
   const offerById = useAppSelector(selectedOffer);
 
-  if (offer === null) {
+  if (offerById === null) {
     <Loader />;
   }
