+++ src/store/api-actions.ts
@@ -2,8 +2,8 @@ import {AxiosInstance} from 'axios';
 import {createAsyncThunk} from '@reduxjs/toolkit';
 import { AppDispatch, State } from '../types/state';
 import { OfferData } from '../types/offers';
-import { OfferNearby } from '../mocks/offers-nearby';
-import { OfferDetailed } from '../mocks/offer';
+import { OfferNearby } from '../types/offers-nearby';
+import { OfferDetailed } from '../types/offer';
 import { setOffers, setOffer, setOffersNearby, requireAuthorization, setError, setOffersDataLoadingStatus } from './action';
 import { saveToken, dropToken } from '../services/token';
 import { APIRoute, AuthorizationStatus } from '../utils/constants';
